refactor(ai): extract Golden Path issue-focus helpers (#13830)
